% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_strata}
\alias{combine_strata}
\title{Function to combine strata via interaction if strata is passed as a vector}
\usage{
combine_strata(x, strata)
}
\description{
Function to combine strata via interaction if strata is passed as a vector
}
\keyword{internal}
