% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_counts}
\alias{get_counts}
\title{To get the n's and response totals with out without strata}
\usage{
get_counts(x, by, strata = 1)
}
\description{
To get the n's and response totals with out without strata
}
\keyword{internal}
