% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_ci.R
\name{ci_prop_clopper_pearson}
\alias{ci_prop_clopper_pearson}
\title{Clopper-Pearson CI}
\usage{
ci_prop_clopper_pearson(x, conf.level = 0.95, data = NULL)
}
\arguments{
\item{x}{(\code{binary}/\code{numeric}/\code{logical})\cr
vector of a binary values, i.e. a logical vector, or numeric with values \code{c(0, 1)}}

\item{conf.level}{(\verb{scalar numeric}) \cr a scalar in (0,1) indicating the confidence level. Default is 0.95}

\item{data}{(\code{data.frame}) \cr Optional data frame containing the variables specified in \code{x} and \code{by}.}
}
\value{
An object containing the following components:

\item{n}{Number of responses}
\item{N}{Total number}
\item{estimate}{The point estimate of the proportion}
\item{conf.low}{Lower bound of the confidence interval}
\item{conf.high}{Upper bound of the confidence interval}
\item{conf.level}{The confidence level used}
\item{method}{Type of method used}
}
\description{
Calculates the Clopper-Pearson interval by calling \code{\link[stats:binom.test]{stats::binom.test()}}.
Also referred to as the \code{exact} method.
}
\details{
\deqn{
\left(1 + \frac{N-n+1}{nF[\frac{\alpha}{2};2n,2(N-n+1)]} \right)^{-1}, \left( 1 + \frac{N-n}{(n+1)F[1-\frac{\alpha}{2};2(n+1),2(N-n)]} \right)^{-1}}
}
